feat(typescript): move typescript variable instances from protractor/globals to protractor
Breaking change for TypeScript:

Instead of importing globals like `browser` from `protractor/globals`, import from `protractor`.

Before:

```ts
import {browser, element} from 'protractor/globals';
import {ElementFinder} from 'protractor';

describe('my app', () => {
  myElement: ElementFinder;

  beforeEach(() => {
    browser.get('example.com');
    myElement = element(by.css('foo'));
  });
});
```

After

```ts
import {browser, element, ElementFinder} from 'protractor';

describe('my app', () => {
  myElement: ElementFinder;

  beforeEach(() => {
    browser.get('example.com');
    myElement = element(by.css('foo'));
  });
});
```

Closes #3564
