--- lib/index.d.ts
@@ -1,7 +1,0 @@
-/// <reference path="../typings/index.d.ts" />
-export {ElementHelper, ProtractorBrowser} from './browser';
-export {Config} from './config';
-export {ElementArrayFinder, ElementFinder} from './element';
-export {ProtractorExpectedConditions} from './expectedConditions';
-export {ProtractorBy} from './locators';
-export {Ptor} from './ptor';
