+++ spec/install/.gitignore
@@ -2,3 +2,4 @@ node_modules
 typings
 conf.js
 typescript_spec.js
+npm-debug.log
