+++ spec/install/typescript_spec.ts
@@ -1,5 +1,5 @@
-import {browser, by, By, element, $, $$, ExpectedConditions, protractor} from 'protractor/globals';
-import {ProtractorBrowser} from 'protractor';
+import {browser, by, By, element, $, $$, ExpectedConditions, protractor, ProtractorBrowser} from 'protractor';
+
 describe('typescript imports', () => {
   it('should have global objects that match the protractor namespace', () => {
     expect(protractor.browser === browser).toBeTruthy();
